Move `CoinjoinAnalyzer` to `BlockchainAnalyzer`
